+++ src/demos/Demos.Collisions.Interactable/Services/App.cs
@@ -23,6 +23,7 @@ internal sealed class App
 	private readonly unsafe WindowHandle* _window;
 	private readonly GlfwInput _glfwInput;
 	private readonly ImGuiController _imGuiController;
+	private readonly AlgorithmParametersWindow _algorithmParametersWindow;
 	private readonly AlgorithmSelectWindow _algorithmSelectWindow;
 	private readonly ScenarioDataWindow _scenarioDataWindow;
 	private readonly SceneRenderer _sceneRenderer;
@@ -39,6 +40,7 @@ internal sealed class App
 		WindowHandle* window,
 		GlfwInput glfwInput,
 		ImGuiController imGuiController,
+		AlgorithmParametersWindow algorithmParametersWindow,
 		AlgorithmSelectWindow algorithmSelectWindow,
 		ScenarioDataWindow scenarioDataWindow,
 		SceneRenderer sceneRenderer,
@@ -49,6 +51,7 @@ internal sealed class App
 		_window = window;
 		_glfwInput = glfwInput;
 		_imGuiController = imGuiController;
+		_algorithmParametersWindow = algorithmParametersWindow;
 		_algorithmSelectWindow = algorithmSelectWindow;
 		_scenarioDataWindow = scenarioDataWindow;
 		_sceneRenderer = sceneRenderer;
@@ -109,11 +112,12 @@ internal sealed class App
 
 		ImGui.DockSpaceOverViewport(0, null, ImGuiDockNodeFlags.PassthruCentralNode);
 
-		_camera.Update(frameTime, true);
+		_camera.Update(frameTime, !ImGui.IsAnyItemHovered());
 		_collisionAlgorithmState.ExecuteAlgorithm();
 
 		_sceneRenderer.Render();
 
+		_algorithmParametersWindow.Render();
 		_algorithmSelectWindow.Render();
 		_scenarioDataWindow.Render();
